web-animations/animation-timeline/animation-timeline-play.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationTimeline.play() method tests</title>
<meta name="assert" content="AnimationTimeline.play() method creates a new AnimationPlayer object">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var animation = new Animation(createDiv(this), [], 2000);
    var player = document.timeline.play(animation);

    assert_not_equals(player, null, 'Player created by AnimationTimeline.play() ' +
        'method should be not null');
}, 'Player created by AnimationTimeline.play() method is not null');

test(function() {
    var player = document.timeline.play(null);

    assert_not_equals(player, null, 'Player created by AnimationTimeline.play(null) ' +
        'method should be not null');
}, 'Player created by AnimationTimeline.play(null) method is not null');

test(function() {
    var player = document.timeline.play();

    assert_not_equals(player, null, 'Player created by AnimationTimeline.play() ' +
        'method without arguments should not be null');
}, 'Player created by AnimationTimeline.play() method without arguments arguments is not null');

test(function() {
    var animation = new Animation(createDiv(this), [], 2000);
    var player = document.timeline.play(animation);

    assert_class_string(player, 'AnimationPlayer', 'Player created by ' +
        'AnimationTimeline.play() method should be AnimationPlayer object instance');
}, 'Test AnimationTimeline.play() method creates a new AnimationPlayer object instance');

test(function() {
    var player = document.timeline.play(null);

    assert_class_string(player, 'AnimationPlayer', 'Player created by ' +
        'AnimationTimeline.play(null) method should be AnimationPlayer object instance');
}, 'Test AnimationTimeline.play(null) method creates a new AnimationPlayer object instance');

test(function() {
    var player = document.timeline.play();

    assert_class_string(player, 'AnimationPlayer', 'Player created by ' +
        'AnimationTimeline.play() method without arguments ' +
        'should be AnimationPlayer object instance');
}, 'Test AnimationTimeline.play() method without arguments creates ' +
    'a new AnimationPlayer object instance');
</script>
</body>
